% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_hidden_encoder.R
\name{build_hidden_encoder}
\alias{build_hidden_encoder}
\title{Build the encoder for a VAE}
\usage{
build_hidden_encoder(
  input_size,
  layers,
  activations = rep("sigmoid", length(layers))
)
}
\arguments{
\item{input_size}{an integer representing the number of items}

\item{layers}{a list of integers giving the size of each hidden layer}

\item{activations}{a list of strings, the same length as layers}
}
\value{
two tensors: the input layer to the VAE and the last hidden layer of the encoder
}
\description{
Build the encoder for a VAE
}
