% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_vae_standard_normal.R
\name{build_vae_standard_normal}
\alias{build_vae_standard_normal}
\title{Build a VAE that fits to a standard N(0,I) latent distribution with independent latent traits}
\usage{
build_vae_standard_normal(num_items, num_skills, Q_matrix,
  enc_hid_arch = c(10), hid_enc_activations = rep("sigmoid",
  length(enc_hid_arch)), output_activation = "sigmoid", kl_weight = 1)
}
\arguments{
\item{num_items}{the number of items on the assessment; also the number of nodes in the input/output layers of the VAE}

\item{num_skills}{the number of skills being evaluated; also the size of the distribution learned by the VAE}

\item{Q_matrix}{a binary, \code{num_skills} by \code{num_items} matrix relating the assessment items with skills}

\item{enc_hid_arch}{a vector detailing the number an size of hidden layers in the encoder}

\item{hid_enc_activations}{a vector specifying the activation function in each hidden layer in the encoder; must be the same length as \code{enc_hid_arch}}

\item{kl_weight}{an optional weight for the KL divergence term in the loss function}
}
\value{
Returns three keras models: the encoder, decoder, and vae.
}
\description{
Build a VAE that fits to a standard N(0,I) latent distribution with independent latent traits
}
\examples{
Q <- matrix(c(1,0,1,1,0,1,1,0), nrow = 2, ncol = 4)
models <- build_vae_standard_normal(4, 2, Q,
          enc_hid_arch = c(6, 3), hid_enc_activation = c('sigmoid', 'relu'),
          output_activation = 'tanh', kl_weight = 0.1)
models <- build_vae_standard_normal(4, 2, Q)
vae <- models[[3]]
}
