% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_matrix.R
\docType{data}
\name{correlation_matrix}
\alias{correlation_matrix}
\title{Simulated latent abilities correlation matrix}
\format{
A data frame with 3 rows and 3 columns
}
\source{
Generated using the python package SciPy
}
\usage{
correlation_matrix
}
\description{
A symmetric positive definite matrix detailing the 
correlations among three latent traits.
}
\keyword{datasets}
