% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_true.R
\docType{data}
\name{diff_true}
\alias{diff_true}
\title{Simulated difficulty parameters}
\format{
A data frame with 30 rows and one column.
Each entry corresponds to the true value of a particular difficulty parameter.
}
\source{
Each entry is sampled uniformly from \code{[-3,3]}.
}
\usage{
diff_true
}
\description{
Difficulty parameters for an exam with 30 items.
}
\keyword{datasets}
