% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Display a message upon loading package}
\usage{
.onLoad(libnam, pkgname)
}
\arguments{
\item{libnam}{the library name}

\item{pkgname}{the package name}
}
\description{
Display a message upon loading package
}
