% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{experimental_standard_loss}
\alias{experimental_standard_loss}
\title{A loss function for a VAE learning a standard normal distribution for use with custom training}
\usage{
experimental_standard_loss(
  z_mean,
  z_log_var,
  kl_weight,
  rec_dim,
  y_true,
  y_pred
)
}
\arguments{
\item{z_mean}{a tensor (layer) representing the mean in the VAE}

\item{z_log_var}{a tensor (layer) representing the log variance in the VAE}

\item{kl_weight}{weight for the KL divergence term}

\item{rec_dim}{the number of nodes in the input/output of the VAE}
}
\description{
A loss function for a VAE learning a standard normal distribution for use with custom training
}
