% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental_train.R
\name{fit_full_cov_model}
\alias{fit_full_cov_model}
\title{This is deprecated and will be removed in future versions. Fit full covariacne models with gradient tape rather than keras}
\usage{
fit_full_cov_model(
  encoder,
  decoder,
  vae,
  mean_vector,
  cov_matrix,
  train_data,
  num_epochs = 10,
  learning_rate = 0.001,
  kl_weight = 1,
  verbose = 1
)
}
\arguments{
\item{encoder}{a compiled keras model}

\item{decoder}{a compiled keras model}

\item{vae}{a compiled keras model}

\item{mean_vector}{the mean values of latent skills}

\item{cov_matrix}{the covariance matrix of latent skills}

\item{train_data}{the data to fit the model on}

\item{num_epochs}{how long to train for}

\item{learning_rate}{the learning rate used for the adam optimizer}

\item{kl_weight}{weight for KL Divergence}

\item{verbose}{level of console output}
}
\description{
This is deprecated and will be removed in future versions. Fit full covariacne models with gradient tape rather than keras
}
