% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental_train.R
\name{fit_standard_model}
\alias{fit_standard_model}
\title{Fit standard normal models with gradient tape rather than keras}
\usage{
fit_standard_model(
  encoder,
  decoder,
  vae,
  train_data,
  num_epochs = 10,
  batch_size = 1L,
  learning_rate = 0.001,
  kl_weight = 1,
  verbose = 1
)
}
\arguments{
\item{encoder}{a compiled keras model}

\item{decoder}{a compiled keras model}

\item{vae}{a compiled keras model}

\item{train_data}{the data to fit the model on}

\item{num_epochs}{how long to train for}

\item{learning_rate}{the learning rate used for the adam optimizer}

\item{kl_weight}{weight for KL Divergence}

\item{verbose}{level of console output}
}
\description{
Fit standard normal models with gradient tape rather than keras
}
