% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameter_estimates.R
\name{get_item_parameter_estimates}
\alias{get_item_parameter_estimates}
\title{Get trainable variables from the decoder, which serve as item parameter estimates.}
\usage{
get_item_parameter_estimates(decoder, model_type = 2)
}
\arguments{
\item{decoder}{a trained keras model}

\item{model_type}{either 1 or 2, specifying a 1 parameter (1PL) or 2 parameter (2PL) model}
}
\value{
a list which contains item parameter estimates. The length of this list is equal to model_type
}
\description{
Get trainable variables from the decoder, which serve as item parameter estimates.
}
\examples{
Q <- matrix(c(1,0,1,1,0,1,1,0), nrow = 2, ncol = 4)
models <- build_vae_standard_normal(4, 2, Q, model_type = 2)
decoder <- models[[2]]
item_parameter_estimates <- get_item_parameter_estimates(decoder, model_type = 2)
}
