% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML2Pvae.R
\docType{package}
\name{ML2Pvae}
\alias{ML2Pvae}
\alias{ML2Pvae-package}
\title{ML2Pvae: A package for creating a VAE whose decoder recovers the parameters of the ML2P model.
The encoder can be used to predict the latent skills based on assessment scores.}
\description{
The ML2Pvae package includes functions which build a VAE with the desired architecture, and
fits the latent skills to either a standard normal (independent) distrubution,
or a multivariate normal distribution with a full covariance matrix.
}
