% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\docType{data}
\name{responses}
\alias{responses}
\title{Response data}
\format{
A data frame with 15 columns and 500 rows.
Entry \code{[j,i]} is 1 if student \code{j} answers item \code{i} correctly, and 0 otherwise.
}
\source{
Generated by sampling from the probability of student success
on a given item according to the ML2P model
}
\usage{
responses
}
\description{
Simulated response sets for 500 students on an exam with 15 items.
}
\keyword{datasets}
