% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construction_functions.R
\name{sampling_correlated}
\alias{sampling_correlated}
\title{A reparameterization in order to sample from the learned multivariate normal distribution of the VAE}
\usage{
sampling_correlated(arg)
}
\arguments{
\item{arg}{a layer of tensors representing the mean and log cholesky transform of the covariance matrix}
}
\description{
A reparameterization in order to sample from the learned multivariate normal distribution of the VAE
}
