% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construction_functions.R
\name{sampling_independent}
\alias{sampling_independent}
\title{A reparameterization in order to sample from the learned standard normal distribution of the VAE}
\usage{
sampling_independent(arg)
}
\arguments{
\item{arg}{a layer of tensors representing the mean and variance}
}
\description{
A reparameterization in order to sample from the learned standard normal distribution of the VAE
}
