% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contruction_functions.R
\name{sampling_normal_full_covariance}
\alias{sampling_normal_full_covariance}
\title{A reparameterization in order to sample from the learned multivariate normal distribution of the VAE}
\usage{
sampling_normal_full_covariance(arg)
}
\arguments{
\item{arg}{a layer of tensors representing the mean and log cholesky transform of the covariance matrix}
}
\description{
A reparameterization in order to sample from the learned multivariate normal distribution of the VAE
}
