% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_script.R
\name{test_fit_model}
\alias{test_fit_model}
\title{TODO documentation
Train the VAE to simulated data and plot some results}
\usage{
test_fit_model(models, Q, num_reps)
}
\arguments{
\item{models}{a list of three models: the encoder, decoder, and vae}

\item{Q}{a binary matrix relating skills and items}

\item{num_reps}{the number of data sets to train on}
}
\description{
TODO documentation
Train the VAE to simulated data and plot some results
}
