% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_script.R
\name{test_normal_full_covariance}
\alias{test_normal_full_covariance}
\title{TODO documentation
This is a script for fitting a normal distribution with a full covariance matrix}
\usage{
test_normal_full_covariance()
}
\description{
TODO documentation
This is a script for fitting a normal distribution with a full covariance matrix
}
