% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta_true.R
\docType{data}
\name{theta_true}
\alias{theta_true}
\title{Simulated ability parameters}
\format{
A data frame with 5000 rows and 3 columns. Each row represents a particular
student's three latent abilities.
}
\source{
Generated by sampling from a 3-dimensional multivariate Gaussian distribution
with mean 0 and covariance matrix \code{correlation_matrix.rda}.
}
\usage{
theta_true
}
\description{
Three correlated ability parameters for 5000 students.
}
\keyword{datasets}
