% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model.R
\name{train_model}
\alias{train_model}
\title{Create a wrapper function to fit models.}
\usage{
train_model(
  model,
  train_data,
  validation_split = 0.15,
  shuffle = FALSE,
  num_epochs = 10,
  verbose = 1,
  batch_size = 1
)
}
\arguments{
\item{model}{the model to be trained}

\item{train_data}{training data}

\item{validation_split}{split percent to use as validation data}

\item{shuffle}{whether or not to shuffle data}

\item{num_epochs}{number of epochs to train for}

\item{verbose}{verbosity levels}

\item{batch_size}{batch size for SGD}
}
\description{
Create a wrapper function to fit models.
}
