% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{vae_loss_normal_full_covariance}
\alias{vae_loss_normal_full_covariance}
\title{A custom loss function for a VAE learning a multivariate normal distribution with a full covariance matrix}
\usage{
vae_loss_normal_full_covariance(z_mean, z_log_cholesky, inv_skill_cov,
  det_skill_cov, skill_mean, kl_weight, rec_dim)
}
\arguments{
\item{z_mean}{a tensor (layer) representing the mean in the VAE}

\item{z_log_cholesky}{a tensor (layer) reshaped into a lower triangular matrix, representing the log cholesky matrix in the VAE}

\item{inv_skill_cov}{a constant tensor matrix of the inverse of the covariance matrix being learned}

\item{det_skill_cov}{a constant tensor scalar representing the determinant of the covariance matrix being learned}

\item{skill_mean}{a constant tensor vector representing the means of the latent skills being learned}

\item{kl_weight}{weight for the KL divergence term}

\item{rec_dim}{the number of nodes in the input/output of the VAE

TODO: change tf$linalg$det() to product of diagonal}
}
\description{
A custom loss function for a VAE learning a multivariate normal distribution with a full covariance matrix
}
