% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{vae_loss_standard_normal}
\alias{vae_loss_standard_normal}
\title{A custom loss function for a VAE learning a standard normal distribution}
\usage{
vae_loss_standard_normal(encoder, kl_weight, rec_dim)
}
\arguments{
\item{encoder}{the encoder model of the VAE, used to obtain z_mean and z_log_var from inputs}

\item{kl_weight}{weight for the KL divergence term}

\item{rec_dim}{the number of nodes in the input/output of the VAE}
}
\description{
A custom loss function for a VAE learning a standard normal distribution
}
