% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_hidden_encoder.R
\name{build_hidden_encoder}
\alias{build_hidden_encoder}
\title{Build the encoder for your VAE}
\usage{
build_hidden_encoder(
  input_size,
  layers,
  activations = rep("sigmoid", length(layers))
)
}
\arguments{
\item{input_size}{an integer}

\item{layers}{a list of integers}

\item{activations}{a list of strings, the same length as layers}
}
\value{
two tensors: the input layer to the VAE and the last hidden layer of the encoder
}
\description{
Build the encoder for your VAE
}
