% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contruction_functions.R
\name{q_constraint}
\alias{q_constraint}
\title{A custom kernel constraint function that restricts weights between the learned distribution and output. Nonzero weights are determined by the Q matrix}
\usage{
q_constraint(Q)
}
\arguments{
\item{Q}{a binary matrixof size \code{num_skills} by \code{num_items}}
}
\description{
A custom kernel constraint function that restricts weights between the learned distribution and output. Nonzero weights are determined by the Q matrix
}
