% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_script.R
\name{test_standard_normal}
\alias{test_standard_normal}
\title{TODO this whole script - mayble load in CDM data
This is a script for testing a VAE that fits a standard normal distribution}
\usage{
test_standard_normal()
}
\description{
TODO this whole script - mayble load in CDM data
This is a script for testing a VAE that fits a standard normal distribution
}
