% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model.R
\name{train_model}
\alias{train_model}
\title{Trains a VAE or autoencoder model. This acts as a wrapper for keras::fit().}
\usage{
train_model(
  model,
  train_data,
  validation_split = 0.15,
  num_epochs = 10,
  batch_size = 1,
  shuffle = FALSE,
  verbose = 1
)
}
\arguments{
\item{model}{the keras model to be trained; this should be the vae returned from \code{build_vae_standard_normal()} or \code{build_vae_normal_full_covariance}}

\item{train_data}{training data; this should be a binary \code{num_students} by \code{num_items} matrix of student responses to an assessment}

\item{validation_split}{split percentage to use as validation data}

\item{num_epochs}{number of epochs to train for}

\item{batch_size}{batch size for mini-batch stochastic gradient descent; default is 1, detailing pure SGD; if a larger batch size is used (e.g. 32), then a larger number of epochs should be set (e.g. 50)}

\item{shuffle}{whether or not to shuffle data}

\item{verbose}{verbosity levels; 0 = silent; 1 = progress bar and epoch message; 2 = epoch message}
}
\value{
a list containing training history; this holds the loss from each epoch which can be plotted
}
\description{
Trains a VAE or autoencoder model. This acts as a wrapper for keras::fit().
}
\examples{
\donttest{
data <- matrix(c(1,1,0,0,1,0,1,1,0,1,1,0), nrow = 3, ncol = 4)
Q <- matrix(c(1,0,1,1,0,1,1,0), nrow = 2, ncol = 4)
models <- build_vae_standard_normal(4, 2, Q)
vae <- models[[3]]
history <- train_model(vae, data, num_epochs = 3, validation_split = 0, verbose = 0)
plot(history)
}
}
