% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{vae_loss_standard_normal}
\alias{vae_loss_standard_normal}
\title{A custom loss function for a VAE learning a standard normal distribution}
\usage{
vae_loss_standard_normal(z_mean, z_log_var, kl_weight, rec_dim)
}
\arguments{
\item{z_mean}{a tensor (layer) representing the mean in the VAE}

\item{z_log_var}{a tensor (layer) representing the log variance in the VAE}

\item{kl_weight}{weight for the KL divergence term}

\item{rec_dim}{the number of nodes in the input/output of the VAE}
}
\description{
A custom loss function for a VAE learning a standard normal distribution
}
